No effective change - code format and tidy

+++ src/main/java/com/avaje/ebeaninternal/server/type/ScalarTypeBigDecimal.java
@@ -83,7 +83,7 @@ public class ScalarTypeBigDecimal extends ScalarTypeBase<BigDecimal> {
   }
 
   public void jsonWrite(JsonGenerator ctx, String name, BigDecimal value) throws IOException {
-    ctx.writeNumberField(name, (BigDecimal) value);
+    ctx.writeNumberField(name, value);
   }
 
 }
